CAMEL-15562: reorganize code to simplify providing out-of-the-box resume strategies
