+++ components/camel-kafka/src/main/java/org/apache/camel/component/kafka/consumer/KafkaAsyncManualCommit.java
@@ -14,7 +14,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.kafka;
+package org.apache.camel.component.kafka.consumer;
+
+import org.apache.camel.component.kafka.KafkaManualCommit;
 
 /**
  * Can be used for forcing async manual offset commit when using Kafka consumer.
