+++ docs/user-manual/modules/ROOT/pages/camel-3x-upgrade-guide-3_15.adoc
@@ -165,6 +165,16 @@ JTA support is moved out of `camel-cdi` to its own `camel-cdi-jta` module.
 
 The `org.apache.camel.cdi.Main` class has moved from `camel-cdi` to `camel-cdi-main` JAR.
 
+=== camel-kafka
+
+The following classes were moved from `org.apache.camel.component.kafka` to `org.apache.camel.component.kafka.consumer`:
+
+* DefaultKafkaManualAsyncCommit
+* DefaultKafkaManualAsyncCommitFactory
+* DefaultKafkaManualSyncCommit
+* DefaultKafkaManualSyncCommitFactory
+* KafkaAsyncManualCommit
+
 === camel-karaf
 
 The `camel-endpointdsl` and `camel-componentdsl` is no longer supported on Apache Karaf.
@@ -183,3 +193,4 @@ Removed the option `startDelaySeconds` as this does not work correctly and cause
 === camel-rabbitmq
 
 The rabbitmq producer has migrated from commons-poll v1 to v2.
+
